kadai ninsyo

+++ app/Http/Controllers/TasksController.php
@@ -14,17 +14,26 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
      
-     // getでtasks/にアクセスされた場合の「一覧表示処理」
-    public function index()
+  
+   public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::all();
-        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
         //　メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+        return view('tasks.index', $data);
     }
+
 
     /**
      * Show the form for creating a new resource.
@@ -59,11 +68,12 @@ class TasksController extends Controller
             'content' => 'required|max:255',
             ]);
             
-        //タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+           
+        ]);
         
         //　トップページへリダイレクトさせる
         return redirect('/');
@@ -82,7 +92,7 @@ class TasksController extends Controller
         //　idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        //　メッセージ詳細ビューでそれを表示
+        //　タスク詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
             ]);
@@ -148,8 +158,11 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        // タスクを削除
-        $task->delete();
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }// タスクを削除
+        
         
         //トップページへリダイレクトさせる
         return redirect('/');
